# internal/api/openapi/swagger.yaml
basePath: /api
definitions:
  api.APIError:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
  api.AccessoryPlan:
    properties:
      active:
        type: boolean
      athlete_id:
        type: integer
      created_at:
        type: string
      day:
        type: integer
      exercise_id:
        type: integer
      exercise_name:
        type: string
      id:
        type: integer
      notes:
        type: string
      sort_order:
        type: integer
      target_rep_max:
        type: integer
      target_rep_min:
        type: integer
      target_sets:
        type: integer
      target_weight:
        type: number
      updated_at:
        type: string
    type: object
...
  api.AccessoryPlanUpdateRequest:
    properties:
      notes:
        type: string
      sort_order:
        type: integer
      target_rep_max:
        type: integer
      target_rep_min:
        type: integer
      target_sets:
        type: integer
      target_weight:
        type: number
    type: object
...
  api.CopyWeekRequest:
    properties:
      source_week:
        example: 1
        type: integer
      target_week:
        example: 2
        type: integer
    type: object
...
  api.NotificationPreferenceRequest:
    properties:
      external:
        type: boolean
      in_app:
        type: boolean
      type:
        example: workout_review
        type: string
    type: object
...
  api.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    REST API for the RepLog workout tracking application.

    Authentication uses session cookies issued by `POST /api/login`
    (or the WebAuthn / token-login endpoints). Cookies are
    `HttpOnly`, `SameSite=Lax`, and (in production) `Secure`.
    The browser sends them automatically on subsequent requests.
  title: RepLog API
  version: 1.0.0
paths:
...
  /athletes/{id}/accessories/{planID}:
    delete:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Plan ID
        in: path
        name: planID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Delete accessory plan
      tags:
      - Athletes
    put:
      consumes:
      - application/json
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Plan ID
        in: path
        name: planID
        required: true
        type: integer
      - description: Plan
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AccessoryPlanUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AccessoryPlan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Update accessory plan
      tags:
      - Athletes
  /athletes/{id}/accessories/{planID}/deactivate:
    post:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Plan ID
        in: path
        name: planID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Deactivate accessory plan
      tags:
      - Athletes
...
  /athletes/{id}/prescription:
    get:
      description: Returns the day's prescribed sets from the athlete's active program
        (if any), with TM-resolved weights.
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Today's prescription
      tags:
      - Athletes
...
  /avatars/upload:
    post:
      consumes:
      - multipart/form-data
      description: Multipart upload; field name is 'file'. Server returns the new
        avatar URL.
      produces:
      - application/json
      responses:
        "200":
          description: 'e.g. {\"avatar_url\": \"/avatars/abc.jpg\"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Upload avatar
      tags:
      - Avatars
...
  /notifications/preferences:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: List notification preferences
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      parameters:
      - description: Preference
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.NotificationPreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Update notification preference for one type
      tags:
      - Notifications
...
  /programs/{id}/copy-week:
    post:
      consumes:
      - application/json
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      - description: Source/target weeks
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CopyWeekRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Copy week of prescribed sets
      tags:
      - Programs
...
swagger: "2.0"
...